Convert KarimsKebabs plugin from kts to kt
